// mobile/src/services/api/schema.d.ts
export interface components {
    schemas: {
// ...
        SessionResponse: {
            /** Avgformscore */
            avgFormScore?: number | null;
            /** Bodyweightkg */
            bodyWeightKg?: number | null;
            /**
             * Createdat
             * Format: date-time
             */
            createdAt: string;
            /** Difficulty */
            difficulty: string;
            /** Durationseconds */
            durationSeconds?: number | null;
            /** Endedat */
            endedAt?: string | null;
            /** Exerciseid */
            exerciseId: string;
            /** Externalloadkg */
            externalLoadKg?: number | null;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Startedat */
            startedAt?: string | null;
            /** Status */
            status: string;
            /** Targetreps */
            targetReps: number;
            /** Targetsets */
            targetSets: number;
            /** Totalreps */
            totalReps: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}